requirements.txt
fastapi
aiogram
uvicorn
python-dotenv
pytest
pytest-asyncio
httpx
pytest-mock